rename wizard module
